requirements.txt
gradio
sentence-transformers
PyPDF2
faiss-cpu
groq
